requirements.txt
ultralytics
opencv-python
numpy
gradio
pillow
easyocr
torch
torchvision
pytesseract
paddlepaddle
paddleocr